% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{add_ui}
\alias{add_ui}
\title{Add UI for plumber to use}
\usage{
add_ui(ui)
}
\arguments{
\item{ui}{A list of that plumber can use to mount
a UI.}
}
\description{
Add UI for plumber to use
}
\details{
\code{\link[=add_ui]{add_ui()}} is used by package like \code{rapidoc}, \code{redoc} and \code{swagger}.
When you load those package, it calls \code{\link[=add_ui]{add_ui()}} to provide a user
interface that can interpret your plumber OpenAPI Specifications.

\code{ui} list expects the following values
\describe{
\item{package}{Name of the package required for the UI.}
\item{name}{Name of the UI.}
\item{index}{A function that returns the HTML content of the landing page of the UI.}
\item{static}{A function that returns the path to the assets (images, javascript, css, fonts) the UI will use.}
}
}
