% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POIs.R
\name{addPeople}
\alias{addPeople}
\title{Inserts people coordinates and traffic to database}
\usage{
addPeople(body)
}
\arguments{
\item{postBodyJson}{Request body with data to insert}
}
\description{
Inserts people coordinates and traffic to database
}
