% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addSerializer}
\alias{addSerializer}
\title{Register a Serializer}
\usage{
addSerializer(name, serializer, verbose = TRUE)
}
\description{
Use \code{\link[=register_serializer]{register_serializer()}} in favor of addSerializer
}
\keyword{internal}
