% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresDBConnector.R
\name{connectToPostgresDB}
\alias{connectToPostgresDB}
\title{Connects to Postgres database}
\usage{
connectToPostgresDB()
}
\value{
Database connection object to perform DB actions with
}
\description{
Connects to Postgres database
}
