% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addSerializer}
\alias{addSerializer}
\alias{getCharacterSet}
\alias{sessionCookie}
\title{Deprecated functions}
\usage{
addSerializer(name, serializer, verbose = TRUE)

getCharacterSet(contentType = NULL)

sessionCookie(
  key,
  name = "plumber",
  expiration = FALSE,
  http = TRUE,
  secure = FALSE,
  sameSite = FALSE
)
}
\description{
Deprecated functions
}
\section{Functions}{
\itemize{
\item \code{addSerializer()}: See \code{\link[=register_serializer]{register_serializer()}}

\item \code{getCharacterSet()}: See \code{\link[=get_character_set]{get_character_set()}}

\item \code{sessionCookie()}: See \code{\link[=session_cookie]{session_cookie()}}

}}
\keyword{internal}
