% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-R6.R
\name{hookable}
\alias{hookable}
\title{Deprecated R6 functions}
\description{
Deprecated R6 functions

Deprecated R6 functions
}
\section{Functions}{
\itemize{
\item \code{hookable}: See \code{\link[=Hookable]{Hookable()}}
}}

\keyword{internal}
\section{Super class}{
\code{\link[plumber:Hookable]{plumber::Hookable}} -> \code{hookable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hookable$new()}}
\item \href{#method-clone}{\code{hookable$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Hookable" data-id="registerHook">}\href{../../plumber/html/Hookable.html#method-registerHook}{\code{plumber::Hookable$registerHook()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Hookable" data-id="registerHooks">}\href{../../plumber/html/Hookable.html#method-registerHooks}{\code{plumber::Hookable$registerHooks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new \code{hookable}. Throws deprecated warning prompting user to use \code{\link{Hookable}}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hookable$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hookable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
