% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_provision}
\alias{do_provision}
\alias{do_configure_https}
\alias{do_deploy_api}
\alias{do_forward}
\alias{do_remove_api}
\alias{do_remove_forward}
\title{DigitalOcean Plumber server}
\usage{
do_provision(...)

do_configure_https(...)

do_deploy_api(...)

do_forward(...)

do_remove_api(...)

do_remove_forward(...)
}
\description{
These methods are now defunct.
Please use the \href{https://github.com/meztez/plumberDeploy}{\code{plumberDeploy}} R package.
}
\keyword{internal}
