% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresDBConnector.R
\name{disconnectFromPostgresDB}
\alias{disconnectFromPostgresDB}
\title{Disconnects from postgres database}
\usage{
disconnectFromPostgresDB(connection)
}
\arguments{
\item{connection}{Postgres connection object}
}
\description{
Disconnects from postgres database
}
