% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_forward}
\alias{do_forward}
\title{Forward Root Requests to an API}
\usage{
do_forward(droplet, path)
}
\arguments{
\item{droplet}{The droplet on which to act. It's expected that this droplet
was provisioned using \code{\link[=do_provision]{do_provision()}}.}

\item{path}{The path to which root requests should be forwarded}
}
\description{
Forward Root Requests to an API
}
