% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_remove_api}
\alias{do_remove_api}
\title{Remove an API from the server}
\usage{
do_remove_api(droplet, path, delete = FALSE)
}
\arguments{
\item{droplet}{The droplet on which to act}

\item{path}{The path/name of the plumber service}

\item{delete}{If \code{TRUE}, will also delete the associated directory
(\code{/var/plumber/whatever}) from the server.}
}
\description{
Removes all services and routing rules associated with a particular service.
Optionally purges the associated API directory from disk.
}
