% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_remove_forward}
\alias{do_remove_forward}
\title{Remove the forwarding rule}
\usage{
do_remove_forward(droplet)
}
\arguments{
\item{droplet}{The droplet on which to act. It's expected that this droplet
was provisioned using \code{\link[=do_provision]{do_provision()}}. See \code{\link[analogsea:droplet]{analogsea::droplet()}} to obtain a reference to a running droplet.}
}
\description{
Removes the forwarding rule from the root path on the server. The server will
no longer forward requests for \code{/} to an application.
}
