% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer.R
\name{endpoint_serializer}
\alias{endpoint_serializer}
\title{Endpoint Serializer with Hooks}
\usage{
endpoint_serializer(serializer, preexec_hook = NULL, postexec_hook = NULL)
}
\arguments{
\item{serializer}{Serializer method to be used.  This method should already have its initialization arguments applied.}

\item{preexec_hook}{Function to be run directly before a \link{PlumberEndpoint} calls its route method.}

\item{postexec_hook}{Function to be run directly after a \link{PlumberEndpoint} calls its route method.}
}
\description{
This method allows serializers to return both \code{preexec} and \code{postexec} hooks in addition to a serializer.
This is useful for graphics device serializers which need a \code{preexec} and \code{postexec} hook to capture the graphics output.
}
\details{
\code{preexec} and \code{postexec} hooks happend directly before and after a route is executed.
These hooks are specific to a single \link{PlumberEndpoint}'s route calculation.
}
\examples{
# The definition of `serializer_device` returns
# * `preexec`, `postexec` hooks
# * a `serializer_content_type` serializer
print(serializer_device)
}
