% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POIs.R
\name{getPOIsInRadius}
\alias{getPOIsInRadius}
\title{Returns points of interest in the specified radius}
\usage{
getPOIsInRadius(lng, lat, radius)
}
\arguments{
\item{lng}{Longitude}

\item{lat}{Latitude}

\item{radius}{Radius}
}
\description{
Returns points of interest in the specified radius
}
