% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POIs.R
\name{getRadiusPolygon}
\alias{getRadiusPolygon}
\title{Returns radius polygon}
\usage{
getRadiusPolygon(lng, lat, radius)
}
\arguments{
\item{lng}{Longitude}

\item{lat}{Latitude}

\item{radius}{Radius}
}
\description{
Returns radius polygon
}
