% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hookable.R
\name{Hookable}
\alias{Hookable}
\title{Hookable}
\description{
Hookable

Hookable
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-registerHook}{\code{Hookable$registerHook()}}
\item \href{#method-registerHooks}{\code{Hookable$registerHooks()}}
\item \href{#method-clone}{\code{Hookable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerHook"></a>}}
\if{latex}{\out{\hypertarget{method-registerHook}{}}}
\subsection{Method \code{registerHook()}}{
Register a hook on a router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hookable$registerHook(stage, handler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stage}}{a character string.}

\item{\code{handler}}{a hook function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerHooks"></a>}}
\if{latex}{\out{\hypertarget{method-registerHooks}{}}}
\subsection{Method \code{registerHooks()}}{
Register hooks on a router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hookable$registerHooks(handlers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{handlers}}{a named list of hook functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hookable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
