% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer.R
\name{hooks_and_serializer}
\alias{hooks_and_serializer}
\title{Hooks and serializer object}
\usage{
hooks_and_serializer(hooks, serializer)
}
\arguments{
\item{hooks}{Hooks to be supplied directly to corresponding \link{PlumberEndpoint} \verb{$registerHooks()} method}

\item{serializer}{Serializer method to be used.  This method should already have it's initialization arugments applied.}
}
\description{
This method allows serializers to return both hooks and a serializer.
This is useful for graphics device serializers which need a \code{preexec} and \code{postexec} hook to capture the graphics output.
}
\examples{
# The definition of `serializer_device` returns
# * `preexec`, `postexec` hooks
# * a `serializer_content_type` serializer
print(serializer_device)
}
