% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{is_plumber}
\alias{is_plumber}
\title{Determine if Plumber object}
\usage{
is_plumber(pr)
}
\arguments{
\item{pr}{Hopefully a \code{\link{Plumber}} object}
}
\value{
Logical value if \code{pr} inherits from \code{\link{Plumber}}
}
\description{
Determine if Plumber object
}
\examples{
is_plumber(Plumber$new()) # TRUE
is_plumber(PlumberEndpoint$new()) # FALSE
}
