% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{mount_interface}
\alias{mount_interface}
\title{Mount Interface UI}
\usage{
mount_interface(interface)
}
\arguments{
\item{interface}{An interface (list) that plumber can use to mount
a UI.}
}
\description{
Mount Interface UI
}
