% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{mountInterface}
\alias{mountInterface}
\title{Mount Interface UI}
\usage{
mountInterface(interface)
}
\arguments{
\item{interface}{An interface (list) that plumber can use to mount
a UI.}
}
\description{
Mount Interface UI
}
