% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi-spec.R
\name{toOpenAPI}
\alias{toOpenAPI}
\alias{fromOpenAPI}
\title{Convert router to/from OpenAPI specifications}
\usage{
toOpenAPI(plumber, output = NULL)

fromOpenAPI(openapi, format = c("json", "yaml"))
}
\arguments{
\item{plumber}{A plumber router or a plumbable file or directory}

\item{output}{An optional filename where to write specifications.}

\item{openapi}{an OpenAPI specifications string, URL or file}

\item{format}{Input file format. Either "json" or "yml". If not
provided, will be guessed from file extension.}
}
\description{
These functions are used to convert between OpenAPI
specifications and plumber file. Plumber only supports a limited
set of OpenAPI specifications.
}
\details{
OpenAPI is a specifications to describe API. More info
can be found at (https://www.openapis.org/)
}
\examples{
pr <- plumber$new()
toOpenAPI(pr)
}
