% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-options.R
\name{optionsPlumber}
\alias{optionsPlumber}
\title{Set plumber options or get non null plumber options}
\usage{
optionsPlumber(
  apiHost = getOption("plumber.apiHost"),
  debug = getOption("plumber.debug"),
  maxRequestSize = getOption("plumber.maxRequestSize"),
  postBody = getOption("plumber.postBody"),
  port = getOption("plumber.port"),
  sharedSecret = getOption("plumber.sharedSecret"),
  swagger.url = getOption("plumber.swagger.url")
)
}
\arguments{
\item{apiHost}{see \code{\link{plumber-options}}}

\item{debug}{see \code{\link{plumber-options}}}

\item{maxRequestSize}{see \code{\link{plumber-options}}}

\item{postBody}{see \code{\link{plumber-options}}}

\item{port}{see \code{\link{plumber-options}}}

\item{sharedSecret}{see \code{\link{plumber-options}}}

\item{swagger.url}{see \code{\link{plumber-options}}}
}
\description{
Set plumber options or get non null plumber options
}
\details{
Sets plumber options. Call without arguments to get current
values.
}
