% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-body.R
\name{parser_query}
\alias{parser_query}
\alias{parser_json}
\alias{parser_text}
\alias{parser_csv}
\alias{parser_tsv}
\alias{parser_yaml}
\alias{parser_read_file}
\alias{parser_rds}
\alias{parser_octet}
\alias{parser_multi}
\title{Plumber Parsers}
\usage{
parser_query()

parser_json(...)

parser_text(parse_fn = identity)

parser_csv(...)

parser_tsv(...)

parser_yaml(...)

parser_read_file(read_fn = readLines)

parser_rds(...)

parser_octet()

parser_multi()
}
\arguments{
\item{...}{parameters supplied to the appropriate internal function}

\item{parse_fn}{function to further decode a text string into an object}

\item{read_fn}{function used to read a the content of a file. Ex: \code{\link[=readRDS]{readRDS()}}}
}
\description{
Parsers are used in Plumber to transform the raw body content received
by a request to the API. Extra parameters may be provided to parser
functions when adding the parser to plumber. This will allow for
non-default behavior.
}
\section{Functions}{
\itemize{
\item \code{parser_query}: Query string parser

\item \code{parser_json}: JSON parser

\item \code{parser_text}: Helper parser to parse plain text

\item \code{parser_csv}: CSV parser

\item \code{parser_tsv}: TSV parser

\item \code{parser_yaml}: YAML parser

\item \code{parser_read_file}: Helper parser that writes the binary post body to a file and reads it back again using \code{read_fn}.
This parser should be used when reading from a file is required.

\item \code{parser_rds}: RDS parser

\item \code{parser_octet}: Octet stream parser

\item \code{parser_multi}: Multi part parser. This parser will then parse each individual body with its respective parser
}}

\examples{
\dontrun{
# Overwrite `text/json` parsing behavior to not allow JSON vectors to be simplified
add_parser("text/json", parser_json(simplifyVector = FALSE))
}
}
