% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{plumb}
\alias{plumb}
\title{Process a Plumber API}
\usage{
plumb(file = NULL, dir = ".")
}
\arguments{
\item{file}{The file to parse as the plumber router definition.}

\item{dir}{The directory containing the \code{plumber.R} file to parse as the
plumber router definition. Alternatively, if an \code{entrypoint.R} file is
found, it will take precedence and be responsible for returning a runnable
router.}
}
\description{
Process a Plumber API
}
\details{
API routers are the core request handler in plumber. A router is responsible for
taking an incoming request, submitting it through the appropriate filters and
eventually to a corresponding endpoint, if one is found.

See the \href{https://www.rplumber.io/articles/programmatic-usage.html}{Programmatic Usage} article for additional
details on the methods available on this object.
}
