% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-options.R
\docType{package}
\name{plumber-package}
\alias{plumber-package}
\alias{_PACKAGE}
\alias{plumber-options}
\title{How to use Plumber options}
\description{
Gives the ability to automatically generate and serve an HTTP API
    from R functions using the annotations in the R documentation around your
    functions.
}
\section{Options}{

There are a number of global options that affect Plumber's behavior. These can
be set globally with \code{options()}.

\describe{
\item{plumber.apiScheme (defaults to \code{http}, or an empty string
when used outside a running router)}{Scheme used to build UI url and server url for
OpenAPI specification.})
\item{plumber.apiHost (defaults to \code{host} defined by \code{run} method, or an empty string
when used outside a running router)}{Host used to build UI url and server url for
OpenAPI specification.})
\item{plumber.apiPort (defaults to \code{port} defined by \code{run} method, or an empty string
when used outside a running router)}{Port used to build UI url and server url for
OpenAPI specification.})
\item{plumber.apiPath (defaults to an empty string)}{Path used to build UI url and server url for
OpenAPI specification.})
\item{plumber.apiURL (defaults to the combination of the above options)}{UI url and server urls for
OpenAPI specification respecting pattern \verb{scheme://host:port/path}.})
\item{plumber.debug (defaults to \code{FALSE})}{Provides more insight into your API errors. Alternatively,
use parameter \code{debug} of plumber router \code{run} method})
\item{plumber.maxRequestSize (defaults to \code{0})}{Maximum length in bytes of request body. Body larger
than maximum are rejected with http error 413. \code{0} means unlimited size.})
\item{plumber.postBody (defaults to \code{TRUE})}{Copy post body content to \code{req$postBody}.
If you do not need it, set to \code{FALSE} for improved API performance.})
\item{plumber.saveFileToDisk (defaults to \code{FALSE})}{By default octetstream parser will return
raw bytes. If you want it to save file sent to an API to disk and return a file path instead, set
to \code{TRUE}. Note that the API has to handle deletion of the file, it is not done automatically.})
\item{plumber.port (defaults to \code{NULL})}{Port Plumber will attempt to use to start http server.
If the port is already in use, server will not be able to start.})
\item{plumber.sharedSecret (defaults to \code{NULL})}{Shared secret used to filter incoming request.
When \code{NULL}, secret is not validated. Otherwise, Plumber compares secret with http header
\code{PLUMBER_SHARED_SECRET}. Failure to match results in http error 400.})
\item{plumber.swagger.url (defaults to \code{NULL})}{Legacy : same function as \code{plumber.ui.callback}})
\item{plumber.ui (defaults to \code{Swagger})}{document})
\item{plumber.ui.callback (defaults to option \code{plumber.swagger.url})}{A function. Called with
a single parameter corresponding to ui url after Plumber server is ready. This can be used
by RStudio to open UI when API is ran for the editor.})
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.rplumber.io (site)
    https://github.com/rstudio/plumber (dev)}
  \item Report bugs at \url{https://github.com/rstudio/plumber/issues}
}

}
\author{
\strong{Maintainer}: Jeff Allen \email{cran@trestletech.com}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com}
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder, funder]
  \item Bruno Tremblay \email{bruno.tremblay@lacapitale.com} [contributor]
  \item Frans van Dunné \email{frans@ixpantia.com} [contributor]
  \item Sebastiaan Vandewoude \email{sebastiaanvandewoude@gmail.com} [contributor]
  \item  SmartBear Software (swagger-ui) [contributor, copyright holder]
}

}
