% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{Plumber}
\alias{Plumber}
\title{Plumber Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb049952070> 
 - attr(*, "name")= chr "Plumber_generator"
}}
\usage{
Plumber
}
\description{
Plumber Router
}
\keyword{datasets}

