% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-step.R
\docType{data}
\name{PlumberEndpoint}
\alias{PlumberEndpoint}
\title{Plumber Endpoint}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberEndpoint
}
\description{
Defines a terminal handler in a PLumber router.
}
\keyword{datasets}
