% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\docType{data}
\name{PlumberProcessor}
\alias{PlumberProcessor}
\title{A plumber processor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberProcessor
}
\description{
A plumber processor
}
\keyword{datasets}

