% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{PlumberRouter}
\alias{PlumberRouter}
\title{Plumber Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f974432b188> 
 - attr(*, "name")= chr "PlumberRouter_generator"
}}
\usage{
PlumberRouter
}
\description{
Plumber Router
}
\keyword{datasets}

