% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-static.R
\docType{data}
\name{PlumberStatic}
\alias{PlumberStatic}
\title{Static file router}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PlumberStatic
}
\description{
Creates a router that is backed by a directory of files on disk.
}
\keyword{datasets}
