% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plumbr.R
\docType{data}
\name{PlumbrRouter}
\alias{PlumbrRouter}
\title{Plumbr Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f84ea44a188> 
 - attr(*, "name")= chr "PlumbrRouter_generator"
}}
\usage{
PlumbrRouter
}
\description{
Plumbr Router
}
\keyword{datasets}

