% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr}
\alias{pr}
\title{Create a new Plumber router}
\usage{
pr(
  file = NULL,
  filters = defaultPlumberFilters,
  envir = new.env(parent = .GlobalEnv)
)
}
\arguments{
\item{file}{path to file to plumb}

\item{filters}{a list of plumber filters}

\item{envir}{an environment to be used as the enclosure for the routers execution}
}
\value{
A new \code{plumber} router
}
\description{
Create a new Plumber router
}
\examples{
\dontrun{
pr() \%>\%
  pr_run()
}

}
