% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_error_handler}
\alias{pr_error_handler}
\title{Set the error handler that is invoked if any filter or endpoint generates an
error}
\usage{
pr_error_handler(pr, fun)
}
\arguments{
\item{pr}{a plumber router}

\item{fun}{a handler function}
}
\value{
The plumber router with a modified error handler
}
\description{
Set the error handler that is invoked if any filter or endpoint generates an
error
}
