% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_object}
\alias{pr_object}
\title{Add an object to Plumber router}
\usage{
pr_object(pr, name, props)
}
\arguments{
\item{pr}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}

\item{name}{A character string. Name of object}

\item{props}{List of properties in the object}
}
\value{
The Plumber router with the defined object added
}
\description{
Objects can be used to provide schema definitions for request and response bodies
}
\examples{
\dontrun{
pr() \%>\%
  pr_object(
     "Car",
     list(
        Id = list(type="int", required=TRUE),
        Manufacturer = list(type="string", required=TRUE),
        Variants = list(type="string", required=FALSE, isArray=TRUE)
     )
  ) \%>\%
  # TODO - unfinished here
  # TODO - work on how responseBody and requestBody will get passed in to these:
  pr_post("/car/update", function() "Hello") \%>\%
  pr_get("/car/<id>", function(id) "Hello") \%>\%
  pr_run()
}

}
