% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_serializer}
\alias{pr_serializer}
\title{Set the default serializer of the router}
\usage{
pr_serializer(pr, serializer)
}
\arguments{
\item{pr}{A plumber router}

\item{serializer}{A serializer function}
}
\value{
The plumber router with the new default serializer
}
\description{
By default, Plumber serializes responses to JSON. This function updates the
default serializer to the function supplied via \code{serializer}
}
