% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_set.R
\name{pr_set_405}
\alias{pr_set_405}
\title{Set the handler that is called when the incoming request uses an unallowed
method}
\usage{
pr_set_405(pr, fun)
}
\arguments{
\item{pr}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}

\item{fun}{A handler function}
}
\value{
The Plumber router with a modified 404 handler
}
\description{
This function allows a custom error message to be returned when a request
cannot uses an unallowed method.
}
\examples{
\dontrun{
handler_405 <- function(req, res) {
  res$status <- 405
  res$body <- "Oops"
}

pr() \%>\%
  pr_get("/hi", function() "Hello") \%>\%
  pr_set_405(handler_405) \%>\%
  pr_run()
}

}
