% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_set.R
\name{pr_set_serializer}
\alias{pr_set_serializer}
\title{Set the default serializer of the router}
\usage{
pr_set_serializer(pr, serializer)
}
\arguments{
\item{pr}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}

\item{serializer}{A serializer function}
}
\value{
The Plumber router with the new default serializer
}
\description{
By default, Plumber serializes responses to JSON. This function updates the
default serializer to the function supplied via \code{serializer}
}
