% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_set.R
\name{pr_set_ui}
\alias{pr_set_ui}
\title{Set the API user interface}
\usage{
pr_set_ui(pr, ui = getOption("plumber.ui", TRUE), ...)
}
\arguments{
\item{pr}{A Plumber router. Note: The supplied Plumber router will also be updated in place as well as returned by the function.}

\item{ui}{a character value or a logical value. Default to \code{plumber.ui} option value.}

\item{...}{Other params to be passed to \code{ui} functions.}
}
\value{
The Plumber router with the new UI settings.
}
\description{
\code{ui} should be either a logical or a character value matching a registered ui.
When \code{TRUE} or a
function, multiple handles will be added to \code{plumber} object. OpenAPI json
file will be served on paths \verb{/openapi.json} and \verb{/swagger.json}. Swagger UI
will be served on paths \verb{/__swagger__/index.html} and \verb{/__swagger__/}. When
using a function, it will receive the Plumber router as the first parameter
and current OpenAPI Specification as the second. This function should return a
list containing OpenAPI Specification.
See \url{http://spec.openapis.org/oas/v3.0.3}
}
\examples{
\dontrun{
## View API using Swagger UI
# Official Website: https://swagger.io/tools/swagger-ui/
# install.packages("swagger")
if (require(swagger)) {
  pr() \%>\%
    pr_set_ui("swagger") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}

## View API using Redoc
# Official Website: https://github.com/Redocly/redoc
# remotes::install_github("https://github.com/meztez/redoc/")
if (require(redoc)) {
  pr() \%>\%
    pr_set_ui("redoc") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}

## View API using RapiDoc
# Official Website: https://github.com/mrin9/RapiDoc
# remotes::install_github("https://github.com/meztez/rapidoc/")
if (require(rapidoc)) {
  pr() \%>\%
    pr_set_ui("rapidoc") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}
}
}
