% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_set.R
\name{pr_set_ui_callback}
\alias{pr_set_ui_callback}
\title{Set the \code{callback} to tell where the API user interface is located}
\usage{
pr_set_ui_callback(
  pr,
  callback = getOption("plumber.ui.callback", getOption("plumber.swagger.url", NULL))
)
}
\arguments{
\item{pr}{A Plumber router. Note: The supplied Plumber router will also be updated in place as well as returned by the function.}

\item{callback}{a callback function for taking action on UI url.}
}
\value{
The Plumber router with the new UI callback setting.
}
\description{
When set, it will be called with a character string corresponding
to the API UI url. This allows RStudio to open \code{swagger} UI when a
Plumber router \code{\link[=pr_run]{pr_run()}} method is executed using default \code{plumber.ui.callback} option.
}
\examples{
\dontrun{
pr() \%>\%
  pr_set_ui_callback(function(url) { message("API location: ", url) }) \%>\%
  pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
  pr_run()
}
}
