% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-cookie.R
\name{randomCookieKey}
\alias{randomCookieKey}
\title{Random cookie key generator}
\usage{
randomCookieKey()
}
\value{
A 64 digit hexadecimal number to be used as a key for cookie encryption.
}
\description{
Random cookie key generator
}
\seealso{
\code{\link{sessionCookie}}
}
