% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rapier.R
\docType{data}
\name{RapierRouter}
\alias{RapierRouter}
\title{Rapier Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fe573a41790> 
 - attr(*, "name")= chr "RapierRouter_generator"
}}
\usage{
RapierRouter
}
\description{
Rapier Router
}
\keyword{datasets}

