% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-httpproblems.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{http_problem_types}
\alias{http_problem}
\alias{bad_request}
\alias{conflict}
\alias{forbidden}
\alias{not_found}
\alias{unauthorized}
\alias{internal_server_error}
\alias{stop_for_http_problem}
\alias{stop_for_bad_request}
\alias{stop_for_conflict}
\alias{stop_for_forbidden}
\alias{stop_for_not_found}
\alias{stop_for_unauthorized}
\alias{stop_for_internal_server_error}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{httpproblems}{\code{\link[httpproblems:http_problem]{bad_request}}, \code{\link[httpproblems:http_problem]{conflict}}, \code{\link[httpproblems:http_problem]{forbidden}}, \code{\link[httpproblems]{http_problem}}, \code{\link[httpproblems]{http_problem_types}}, \code{\link[httpproblems:http_problem]{internal_server_error}}, \code{\link[httpproblems:http_problem]{not_found}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_bad_request}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_conflict}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_forbidden}}, \code{\link[httpproblems]{stop_for_http_problem}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_internal_server_error}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_not_found}}, \code{\link[httpproblems:stop_for_http_problem]{stop_for_unauthorized}}, \code{\link[httpproblems:http_problem]{unauthorized}}}
}}

