% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-cookie.R
\name{sessionCookie}
\alias{sessionCookie}
\title{Store session data in encrypted cookies.}
\usage{
sessionCookie(key = randomCookieKey(), name = "plumber", ...)
}
\arguments{
\item{key}{The secret key to use. This must be consistent across all sessions
where you want to save/restore encrypted cookies. It should be a long and
complex character string to bolster security. Raw vectors with a length of
32 or more may also be used.}

\item{name}{The name of the cookie in the user's browser.}

\item{...}{Arguments passed on to the \code{response$setCookie} call to,
for instance, set the cookie's expiration.}
}
\description{
\code{plumber} is using the crypto R package \code{sodium}, which binds to
\code{libsodium}, "a modern, easy-to-use software library for encryption".
Using \code{sodium}, \code{req$session} information is be preserved between
requests in encrypted cookies.
}
\details{
Currently, if no \code{key} value is provided, encrypted browser cookies can be valid
until the \code{plumber} server is restarted.  Once restarted,  the new
\code{plumber} server will create a new encryption key and all prior cookies
will silently fail to parse, losing all \code{req$session} information.

If a consistent \code{key} is provided each time a \code{plumber} server is
launched, users will maintain \code{req$session} information between server resets.
}
\section{Storing secure key}{

While it is very quick to get started with user session cookies using
\code{plumber}, please use exercise precaution when storing secure key information.

Please: \itemize{
\item Do NOT store keys in storage code.
\item Do NOT store keys which can be accessed by everyone.
\item Do NOT store keys which can be queried by everyone.
}

Instead, please: \itemize{
\item Use a key management system or password application.
\item (or) Store it on a disk that only you have access to.
Such as modifying the file permissions to "user read only" (\code{chmod 400 myfile.txt}) to your file from being modified and prevent others from reading your file.
}
}

