% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_api.R
\name{validate_api}
\alias{validate_api}
\title{Validate OpenAPI Spec}
\usage{
validate_api(pr)
}
\arguments{
\item{pr}{A Plumber API}
}
\description{
Validate an OpenAPI Spec using \href{https://github.com/APIDevTools/swagger-cli}{Swagger CLI} which calls \href{https://github.com/APIDevTools/swagger-parser}{Swagger Parser}.
}
\details{
If the api is deemed invalid, an error will be thrown.

This function is VERY experimental and may be altered, changed, or removed in the future.
}
