% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer.R
\name{addSerializer}
\alias{addSerializer}
\title{Add a Serializer}
\usage{
addSerializer(name, serializer)
}
\arguments{
\item{name}{The name of the serializer (character string)}

\item{serializer}{The serializer to be added.}
}
\description{
A serializer is responsible for translating a generated R value into output
that a remote user can understand. For instance, the \code{jsonSerializer}
serializes R objects into JSON before returning them to the user. The list of
available serializers in plumber is global.
}

