% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital-ocean.R
\name{do_provision}
\alias{do_provision}
\title{Provision a DigitalOcean plumber server}
\usage{
do_provision(dropletId, ...)
}
\arguments{
\item{dropletId}{The numeric identifier of the DigitalOcean server that you want to provision (see \code{\link[analogsea:droplets]{analogsea::droplets()}}). If empty, a new DigitalOcean server will be created.}

\item{...}{Arguments passed into the \code{\link[analogsea:droplet_create]{analogsea::droplet_create()}} function.}
}
\description{
Create (if required), install the necessary prerequisites, and
deploy a sample plumber application on a DigitalOcean virtual machine.
}
