% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi-from-to.R
\name{to_openapi}
\alias{to_openapi}
\alias{from_openapi}
\title{Convert router to/from OpenAPI specifications}
\usage{
to_openapi(plumber, output = NULL)

from_openapi(openapi, format = c("json", "yaml"))
}
\arguments{
\item{output}{An optional filename where to write OpenAPI Specifications.}

\item{openapi}{an OpenAPI specifications string, URL or file}

\item{format}{Input file format. Either "json" or "yml". If not
provided, will be guessed from file extension.}

\item{pr}{A plumber router or a plumbable file or directory.}
}
\description{
These functions are used to convert between OpenAPI
Specifications and plumber file. Plumber only supports a limited
set of OpenAPI Specifications.
}
\details{
OpenAPI Specifications are used to document APIs. More info
can be found at (https://www.openapis.org/)
}
\examples{
pr <- plumber$new()
to_openapi(pr)
}
