% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-options.R
\name{options_plumber}
\alias{options_plumber}
\title{Set plumber options}
\usage{
options_plumber(
  apiHost = getOption("plumber.apiHost"),
  debug = getOption("plumber.debug"),
  maxRequestSize = getOption("plumber.maxRequestSize"),
  postBody = getOption("plumber.postBody"),
  port = getOption("plumber.port"),
  sharedSecret = getOption("plumber.sharedSecret"),
  swagger.url = getOption("plumber.swagger.url")
)
}
\arguments{
\item{apiHost}{see \link{plumber-options}}

\item{debug}{see \link{plumber-options}}

\item{maxRequestSize}{see \link{plumber-options}}

\item{postBody}{see \link{plumber-options}}

\item{port}{see \link{plumber-options}}

\item{sharedSecret}{see \link{plumber-options}}

\item{swagger.url}{see \link{plumber-options}}
}
\description{
Set plumber options
}
\details{
Sets plumber options. Call without arguments to get current
values.
}
