% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-parsers.R
\name{parsers}
\alias{parsers}
\alias{parser_json}
\alias{parser_query}
\alias{parser_text}
\alias{parser_rds}
\alias{parser_multi}
\alias{parser_octet}
\title{Plumber Parsers}
\usage{
parser_json(...)

parser_query(...)

parser_text(...)

parser_rds(...)

parser_multi(...)

parser_octet(...)
}
\arguments{
\item{...}{Raw values and headers are passed there.}
}
\description{
Parsers are used in Plumber to transform the raw body content received
by a request to the API.
}
