% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{plumb_api}
\alias{plumb_api}
\alias{available_apis}
\title{Process a Package's Plumber API}
\usage{
plumb_api(package = NULL, name = NULL)

available_apis(package = NULL)
}
\arguments{
\item{package}{Package to inspect}

\item{name}{Name of the package folder to \code{\link[=plumb]{plumb()}}.}
}
\description{
So that packages can ship multiple plumber routers, users should store their Plumber APIs
in the \code{inst} subfolder \code{plumber} (\code{./inst/plumber/API_1/plumber.R}).
}
\details{
To view all available Plumber APIs across all packages, please call \code{available_apis()}.
A \code{package} value may be provided to only display a particular package's Plumber APIs.
}
\section{Functions}{
\itemize{
\item \code{plumb_api}: \code{\link[=plumb]{plumb()}}s a package's Plumber API. Returns a \code{\link{plumber}} router object

\item \code{available_apis}: Displays all available package Plumber APIs. Returns a \code{data.frame} of \code{package} and \code{name} information.
}}

