% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{plumb}
\alias{plumb}
\alias{plumber}
\title{Plumber Router}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
plumb(file, dir = ".", envir)

plumber
}
\arguments{
\item{file}{The file to parse as the plumber router definition}

\item{dir}{The directory containing the \code{plumber.R} file to parse as the
plumber router definition. Alternatively, if an \code{entrypoint.R} file is
found, it will take precedence and be responsible for returning a runnable
Plumber router.}

\item{envir}{The environment in which to evaluate all expressions, source any
input files, and execute all filters/endpoints. If unspecified, a new
environment will be created for you. Note that this setting has no effect if
you create your own router using an \code{entrypoint.R} file.}
}
\description{
Routers are the core request handler in plumber. A router is responsible for
taking an incoming request, submitting it through the appropriate filters and
eventually to a corresponding endpoint, if one is found.
}
\details{
See \url{http://plumber.trestletech.com/docs/programmatic/} for additional
details on the methods available on this object.
}
\keyword{datasets}
