% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{plumber}
\alias{plumber}
\title{Plumber Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb73c7da120> 
 - attr(*, "name")= chr "plumber_generator"
}}
\usage{
plumber
}
\description{
Plumber Router
}
\keyword{datasets}

