% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{plumber}
\alias{plumb}
\alias{plumber}
\title{Plumber Router}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f952bf5f800> 
 - attr(*, "name")= chr "plumber_generator"
}}
\usage{
plumber

plumb(file)
}
\description{
Routers are the core request handler in plumber. A router is responsible for
taking an incoming request, submitting it through the appropriate filters and
eventually to a corresponding endpoint, if one is found.
}
\details{
See \url{http://plumber.trestletech.com/docs/programmatic/} for additional
details on the methods available on this object.
}
\keyword{datasets}

