% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{plumb}
\alias{plumb}
\alias{plumber}
\title{Plumber Router}
\usage{
plumb(file = NULL, dir = ".")
}
\arguments{
\item{file}{The file to parse as the plumber router definition.}

\item{dir}{The directory containing the \code{plumber.R} file to parse as the
plumber router definition. Alternatively, if an \code{entrypoint.R} file is
found, it will take precedence and be responsible for returning a runnable
router.}
}
\description{
Plumber Router
}
\details{
Routers are the core request handler in plumber. A router is responsible for
taking an incoming request, submitting it through the appropriate filters and
eventually to a corresponding endpoint, if one is found.

See \url{http://www.rplumber.io/docs/programmatic/} for additional
details on the methods available on this object.
}
\examples{

## ------------------------------------------------
## Method `plumber$run`
## ------------------------------------------------

\dontrun{
pr <- plumber$new
swagger <- function(pr_, spec) {
  spec$servers[[1]]$description <- "MyCustomAPISpec"
  spec
}
pr$run(swagger = swagger)
}

## ------------------------------------------------
## Method `plumber$mount`
## ------------------------------------------------

\dontrun{
root <- plumber$new()

users <- plumber$new("users.R")
root$mount("/users", users)

products <- plumber$new("products.R")
root$mount("/products", products)
}

## ------------------------------------------------
## Method `plumber$registerHook`
## ------------------------------------------------

\dontrun{
pr <- plumber$new()
pr$registerHook("preroute", function(req){
  cat("Routing a request for", req$PATH_INFO, "...\n")
})
pr$registerHooks(list(
  preserialize=function(req, value){
    print("About to serialize this value:")
    print(value)

    # Must return the value since we took one in. Here we're not choosing
    # to mutate it, but we could.
    value
  },
  postserialize=function(res){
    print("We serialized the value as:")
    print(res$body)
  }
))

pr$handle("GET", "/", function(){ 123 })
}

## ------------------------------------------------
## Method `plumber$handle`
## ------------------------------------------------

\dontrun{
pr <- plumber$new()
pr$handle("GET", "/", function(){
  "<html><h1>Programmatic Plumber!</h1></html>"
}, serializer=plumber::serializer_html())
}

## ------------------------------------------------
## Method `plumber$set404Handler`
## ------------------------------------------------

\dontrun{
pr <- plumber$new()
pr$set404Handler(function(req, res) {cat(req$PATH_INFO)})
}

## ------------------------------------------------
## Method `plumber$setErrorHandler`
## ------------------------------------------------

\dontrun{
pr <- plumber$new()
pr$setErrorHandler(function(req, res) {cat(res$body)})
}
}
\section{Super class}{
\code{\link[plumber:hookable]{plumber::hookable}} -> \code{plumber}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{endpoints}}{plumber router endpoints read-only}

\item{\code{filters}}{plumber router filters read-only}

\item{\code{mounts}}{plumber router mounts read-only}

\item{\code{environment}}{plumber router environment read-only}

\item{\code{routes}}{plumber router routes read-only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{plumber$new()}}
\item \href{#method-run}{\code{plumber$run()}}
\item \href{#method-mount}{\code{plumber$mount()}}
\item \href{#method-registerHook}{\code{plumber$registerHook()}}
\item \href{#method-handle}{\code{plumber$handle()}}
\item \href{#method-print}{\code{plumber$print()}}
\item \href{#method-serve}{\code{plumber$serve()}}
\item \href{#method-route}{\code{plumber$route()}}
\item \href{#method-call}{\code{plumber$call()}}
\item \href{#method-onHeaders}{\code{plumber$onHeaders()}}
\item \href{#method-onWSOpen}{\code{plumber$onWSOpen()}}
\item \href{#method-setSerializer}{\code{plumber$setSerializer()}}
\item \href{#method-set404Handler}{\code{plumber$set404Handler()}}
\item \href{#method-setErrorHandler}{\code{plumber$setErrorHandler()}}
\item \href{#method-filter}{\code{plumber$filter()}}
\item \href{#method-apiSpec}{\code{plumber$apiSpec()}}
\item \href{#method-openAPIFile}{\code{plumber$openAPIFile()}}
\item \href{#method-swaggerFile}{\code{plumber$swaggerFile()}}
\item \href{#method-addEndpoint}{\code{plumber$addEndpoint()}}
\item \href{#method-addAssets}{\code{plumber$addAssets()}}
\item \href{#method-addFilter}{\code{plumber$addFilter()}}
\item \href{#method-addGlobalProcessor}{\code{plumber$addGlobalProcessor()}}
\item \href{#method-clone}{\code{plumber$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="hookable" data-id="registerHooks">}\href{../../plumber/html/hookable.html#method-registerHooks}{\code{plumber::hookable$registerHooks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{plumber} router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$new(file = NULL, filters = defaultPlumberFilters, envir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{path to file to plumb}

\item{\code{filters}}{a list of plumber filters}

\item{\code{envir}}{an environment to be used as the enclosure for the routers execution}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{plumber} router
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Start a server using \code{plumber} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$run(
  host = "127.0.0.1",
  port = getOption("plumber.port"),
  swagger = interactive(),
  debug = interactive(),
  swaggerCallback = getOption("plumber.swagger.url", NULL)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{a string that is a valid IPv4 or IPv6 address that is owned by
this server, which the application will listen on. "0.0.0.0" represents
all IPv4 addresses and "::/0" represents all IPv6 addresses.}

\item{\code{port}}{a number or integer that indicates the server port that should
be listened on. Note that on most Unix-like systems including Linux and
Mac OS X, port numbers smaller than 1025 require root privileges.}

\item{\code{swagger}}{a function that enhances the existing OpenAPI Specification.}

\item{\code{debug}}{\code{TRUE} provides more insight into your API errors.}

\item{\code{swaggerCallback}}{a callback function for taking action on the url for swagger page.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\code{port} does not need to be explicitly assigned.

\code{swagger} should be either a logial or a function . When \code{TRUE} or a
function, multiple handles will be added to \code{plumber} object. OpenAPI json
file will be served on paths \verb{/openapi.json} and \verb{/swagger.json}. Swagger UI
will be served on paths \verb{/__swagger__/index.html} and \verb{/__swagger__/}. When
using a function, it will receive the plumber router as the first parameter
and current OpenAPI Specification as the second. This function should return a
list containing OpenAPI Specification.
See \url{http://spec.openapis.org/oas/v3.0.3}

\code{swaggerCallback} When set, it will be called with a character string corresponding
to the swagger UI url. It allows RStudio to open swagger UI when plumber router
run method is executed using default \code{plumber.swagger.url} option.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
pr <- plumber$new
swagger <- function(pr_, spec) {
  spec$servers[[1]]$description <- "MyCustomAPISpec"
  spec
}
pr$run(swagger = swagger)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mount"></a>}}
\if{latex}{\out{\hypertarget{method-mount}{}}}
\subsection{Method \code{mount()}}{
Mount a plumber router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$mount(path, router)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{a character string. Where to mount router.}

\item{\code{router}}{a plumber router. Router to be mounted.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Plumber routers can be “nested” by mounting one into another
using the \code{mount()} method. This allows you to compartmentalize your API
by paths which is a great technique for decomposing large APIs into smaller files.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
root <- plumber$new()

users <- plumber$new("users.R")
root$mount("/users", users)

products <- plumber$new("products.R")
root$mount("/products", products)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerHook"></a>}}
\if{latex}{\out{\hypertarget{method-registerHook}{}}}
\subsection{Method \code{registerHook()}}{
Register a hook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$registerHook(
  stage = c("preroute", "postroute", "preserialize", "postserialize", "exit"),
  handler
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stage}}{a character string. Point in the lifecycle of a request.}

\item{\code{handler}}{a hook function.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Plumber routers support the notion of "hooks" that can be registered
to execute some code at a particular point in the lifecycle of a request.
Plumber routers currently support four hooks:
\enumerate{
\item \code{preroute(data, req, res)}
\item \code{postroute(data, req, res, value)}
\item \code{preserialize(data, req, res, value)}
\item \code{postserialize(data, req, res, value)}
}
In all of the above you have access to a disposable environment in the \code{data}
parameter that is created as a temporary data store for each request. Hooks
can store temporary data in these hooks that can be reused by other hooks
processing this same request.

One feature when defining hooks in Plumber routers is the ability to modify
the returned value. The convention for such hooks is: any function that accepts
a parameter named \code{value} is expected to return the new value. This could
be an unmodified version of the value that was passed in, or it could be a
mutated value. But in either case, if your hook accepts a parameter
named \code{value}, whatever your hook returns will be used as the new value
for the response.

You can add hooks using the \code{registerHook} method, or you can add multiple
hooks at once using the \code{registerHooks} method which takes a name list in
which the names are the names of the hooks, and the values are the
handlers themselves.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
pr <- plumber$new()
pr$registerHook("preroute", function(req){
  cat("Routing a request for", req$PATH_INFO, "...\n")
})
pr$registerHooks(list(
  preserialize=function(req, value){
    print("About to serialize this value:")
    print(value)

    # Must return the value since we took one in. Here we're not choosing
    # to mutate it, but we could.
    value
  },
  postserialize=function(res){
    print("We serialized the value as:")
    print(res$body)
  }
))

pr$handle("GET", "/", function(){ 123 })
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handle"></a>}}
\if{latex}{\out{\hypertarget{method-handle}{}}}
\subsection{Method \code{handle()}}{
Define endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$handle(methods, path, handler, preempt, serializer, endpoint, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{methods}}{a character string. http method.}

\item{\code{path}}{a character string. Api endpoints}

\item{\code{handler}}{a handler function.}

\item{\code{preempt}}{a preempt function.}

\item{\code{serializer}}{a serializer function.}

\item{\code{endpoint}}{a \code{PlumberEndpoint} object.}

\item{\code{...}}{additional arguments for \code{PlumberEndpoint} creation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The “handler” functions that you define in these handle calls
are identical to the code you would have defined in your plumber.R file
if you were using annotations to define your API. The handle() method
takes additional arguments that allow you to control nuanced behavior
of the endpoint like which filter it might preempt or which serializer
it should use.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
pr <- plumber$new()
pr$handle("GET", "/", function(){
  "<html><h1>Programmatic Plumber!</h1></html>"
}, serializer=plumber::serializer_html())
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print reprensation of plumber router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$print(prefix = "", topLevel = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{a character string. Prefix to append to representation.}

\item{\code{topLevel}}{a logical value. When method executed on top level
router, set to \code{TRUE}.}

\item{\code{...}}{additional arguments for recursive calls}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A terminal friendly represention of a plumber router.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serve"></a>}}
\if{latex}{\out{\hypertarget{method-serve}{}}}
\subsection{Method \code{serve()}}{
Serve a request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$serve(req, res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{request object}

\item{\code{res}}{response object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-route"></a>}}
\if{latex}{\out{\hypertarget{method-route}{}}}
\subsection{Method \code{route()}}{
Route a request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$route(req, res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{request object}

\item{\code{res}}{response object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\if{latex}{\out{\hypertarget{method-call}{}}}
\subsection{Method \code{call()}}{
httpuv interface call function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$call(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{request object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
required for httpuv interface
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-onHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-onHeaders}{}}}
\subsection{Method \code{onHeaders()}}{
httpuv interface onHeaders function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$onHeaders(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{request object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
required for httpuv interface
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-onWSOpen"></a>}}
\if{latex}{\out{\hypertarget{method-onWSOpen}{}}}
\subsection{Method \code{onWSOpen()}}{
httpuv interface onWSOpen function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$onWSOpen(ws)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ws}}{WebSocket object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
required for httpuv interface
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSerializer"></a>}}
\if{latex}{\out{\hypertarget{method-setSerializer}{}}}
\subsection{Method \code{setSerializer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$setSerializer(serializer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{serializer}}{a serializer function}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the default serializer of the router.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set404Handler"></a>}}
\if{latex}{\out{\hypertarget{method-set404Handler}{}}}
\subsection{Method \code{set404Handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$set404Handler(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{a handler function.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the handler that gets called if an
incoming request can’t be served by any filter, endpoint, or sub-router.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
pr <- plumber$new()
pr$set404Handler(function(req, res) {cat(req$PATH_INFO)})
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setErrorHandler"></a>}}
\if{latex}{\out{\hypertarget{method-setErrorHandler}{}}}
\subsection{Method \code{setErrorHandler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$setErrorHandler(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{a handler function.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the error handler which gets invoked if any filter or
endpoint generates an error.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
pr <- plumber$new()
pr$setErrorHandler(function(req, res) {cat(res$body)})
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\if{latex}{\out{\hypertarget{method-filter}{}}}
\subsection{Method \code{filter()}}{
Add a filter to plumber router
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$filter(name, expr, serializer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a character string. Name of filter}

\item{\code{expr}}{an expr that resolve to a filter function or a filter function}

\item{\code{serializer}}{a serializer function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apiSpec"></a>}}
\if{latex}{\out{\hypertarget{method-apiSpec}{}}}
\subsection{Method \code{apiSpec()}}{
Retrieve openAPI file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$apiSpec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openAPIFile"></a>}}
\if{latex}{\out{\hypertarget{method-openAPIFile}{}}}
\subsection{Method \code{openAPIFile()}}{
Retrieve openAPI file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$openAPIFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-swaggerFile"></a>}}
\if{latex}{\out{\hypertarget{method-swaggerFile}{}}}
\subsection{Method \code{swaggerFile()}}{
Retrieve openAPI file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$swaggerFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addEndpoint"></a>}}
\if{latex}{\out{\hypertarget{method-addEndpoint}{}}}
\subsection{Method \code{addEndpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$addEndpoint(
  verbs,
  path,
  expr,
  serializer,
  processors,
  preempt = NULL,
  params = NULL,
  comments
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbs}}{verbs}

\item{\code{path}}{path}

\item{\code{expr}}{expr}

\item{\code{serializer}}{serializer}

\item{\code{processors}}{processors}

\item{\code{preempt}}{preempt}

\item{\code{params}}{params}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
addEndpoint has been deprecated in v0.4.0 and will be removed in a coming release. Please use \code{handle()} instead.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addAssets"></a>}}
\if{latex}{\out{\hypertarget{method-addAssets}{}}}
\subsection{Method \code{addAssets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$addAssets(dir, path = "/public", options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{dir}

\item{\code{path}}{path}

\item{\code{options}}{options}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
addAssets has been deprecated in v0.4.0 and will be removed in a coming release. Please use \code{mount} and \code{PlumberStatic$new()} instead.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addFilter"></a>}}
\if{latex}{\out{\hypertarget{method-addFilter}{}}}
\subsection{Method \code{addFilter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$addFilter(name, expr, serializer, processors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{expr}}{expr}

\item{\code{serializer}}{serializer}

\item{\code{processors}}{processors}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
addFilter has been deprecated in v0.4.0 and will be removed in a coming release. Please use \code{filter} instead.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addGlobalProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-addGlobalProcessor}{}}}
\subsection{Method \code{addGlobalProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$addGlobalProcessor(proc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proc}}{proc}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
addGlobalProcessor has been deprecated in v0.4.0 and will be removed in a coming release. Please use \code{registerHook}(s) instead.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plumber$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
