% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\docType{data}
\name{plumber}
\alias{plumb}
\alias{plumber}
\title{Plumber Router}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
plumber

plumb(file)
}
\arguments{
\item{file}{The file to parse as the plumber router definition}
}
\description{
Routers are the core request handler in plumber. A router is responsible for
taking an incoming request, submitting it through the appropriate filters and
eventually to a corresponding endpoint, if one is found.
}
\details{
See \url{http://plumber.trestletech.com/docs/programmatic/} for additional
details on the methods available on this object.
}
\keyword{datasets}

