% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-step.R
\name{PlumberStep}
\alias{PlumberStep}
\title{plumber step R6 class}
\description{
an object representing a step in the lifecycle of the treatment
of a request by a plumber router.
}
\section{Super class}{
\code{\link[plumber:hookable]{plumber::hookable}} -> \code{PlumberStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lines}}{lines from step block}

\item{\code{serializer}}{step serializer function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlumberStep$new()}}
\item \href{#method-exec}{\code{PlumberStep$exec()}}
\item \href{#method-registerHook}{\code{PlumberStep$registerHook()}}
\item \href{#method-clone}{\code{PlumberStep$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="hookable" data-id="registerHooks">}\href{../../plumber/html/hookable.html#method-registerHooks}{\code{plumber::hookable$registerHooks()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{\link[=PlumberStep]{PlumberStep()}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumberStep$new(expr, envir, lines, serializer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{step expr}

\item{\code{envir}}{step environment}

\item{\code{lines}}{step block}

\item{\code{serializer}}{step serializer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PlumberStep} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exec"></a>}}
\if{latex}{\out{\hypertarget{method-exec}{}}}
\subsection{Method \code{exec()}}{
step execution function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumberStep$exec(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments for step execution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerHook"></a>}}
\if{latex}{\out{\hypertarget{method-registerHook}{}}}
\subsection{Method \code{registerHook()}}{
step hook registration method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumberStep$registerHook(stage = c("preexec", "postexec"), handler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stage}}{a character string.}

\item{\code{handler}}{a step handler function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlumberStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
