% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_run}
\alias{pr_run}
\title{Start a server using \code{plumber} object}
\usage{
pr_run(pr, host = "127.0.0.1", port = getOption("plumber.port"))
}
\arguments{
\item{pr}{A Plumber router}

\item{host}{A string that is a valid IPv4 or IPv6 address that is owned by
this server, which the application will listen on. "0.0.0.0" represents
all IPv4 addresses and "::/0" represents all IPv6 addresses.}

\item{port}{A number or integer that indicates the server port that should
be listened on. Note that on most Unix-like systems including Linux and
Mac OS X, port numbers smaller than 1025 require root privileges.}
}
\description{
\code{port} does not need to be explicitly assigned.
}
\details{
\code{swagger} should be either a logical or a function . When \code{TRUE} or a
function, multiple handles will be added to \code{plumber} object. OpenAPI json
file will be served on paths \verb{/openapi.json} and \verb{/swagger.json}. Swagger UI
will be served on paths \verb{/__swagger__/index.html} and \verb{/__swagger__/}. When
using a function, it will receive the Plumber router as the first parameter
and current OpenAPI Specification as the second. This function should return a
list containing OpenAPI Specification.
See \url{http://spec.openapis.org/oas/v3.0.3}

\code{swaggerCallback} When set, it will be called with a character string corresponding
to the swagger UI url. It allows RStudio to open swagger UI when Plumber router
run method is executed using default \code{plumber.swagger.url} option.
}
\examples{
\dontrun{
pr() \%>\%
  pr_run()

pr() \%>\%
  pr_run(port = 5762, debug = TRUE)
}

}
