% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_set.R
\name{pr_set_ui}
\alias{pr_set_ui}
\title{Set the API user interface}
\usage{
pr_set_ui(
  pr,
  ui = getOption("plumber.ui", TRUE),
  callback = getOption("plumber.ui.callback", getOption("plumber.swagger.url", NULL)),
  ...
)
}
\arguments{
\item{pr}{A Plumber router}

\item{ui}{a character value or a logical value. Default to \code{plumber.ui} option value.}

\item{callback}{a callback function for taking action on UI url.}

\item{...}{Other params to be passed to \code{ui} functions.}
}
\value{
The Plumber router with the new UI settings.
}
\description{
Set the API user interface
}
\examples{
\dontrun{
## View API using Swagger UI
# Official Website: https://swagger.io/tools/swagger-ui/
# install.packages("swagger")
if (require(swagger)) {
  pr() \%>\%
    pr_set_ui("swagger") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}

## View API using Redoc
# Official Website: https://github.com/Redocly/redoc
# remotes::install_github("https://github.com/meztez/redoc/")
if (require(redoc)) {
  pr() \%>\%
    pr_set_ui("redoc") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}

## View API using RapiDoc
# Official Website: https://github.com/mrin9/RapiDoc
# remotes::install_github("https://github.com/meztez/rapidoc/")
if (require(rapidoc)) {
  pr() \%>\%
    pr_set_ui("rapidoc") \%>\%
    pr_get("/plus/<a:int>/<b:int>", function(a, b) { a + b }) \%>\%
    pr_run()
}
}
}
