% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr_static}
\alias{pr_static}
\title{Add a static route to the \code{plumber} object}
\usage{
pr_static(pr, path, direc)
}
\arguments{
\item{pr}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}

\item{path}{The mounted path location of the static folder}

\item{direc}{The local folder to be served statically}
}
\description{
Add a static route to the \code{plumber} object
}
\examples{
\dontrun{
pr() \%>\%
  pr_static("/path", "./my_folder/location") \%>\%
  pr_run()
}

}
