% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-cookie.R
\name{randomCookieKey}
\alias{randomCookieKey}
\title{Random cookie key generator}
\usage{
randomCookieKey()
}
\value{
A 64 digit hexadecimal string to be used as a key for cookie encryption.
}
\description{
Uses a cryptographically secure pseudorandom number generator from \code{\link[sodium:helpers]{sodium::helpers()}} to generate a 64 digit hexadecimal string.  \href{https://github.com/jeroen/sodium}{'sodium'} wraps around \href{https://download.libsodium.org/doc/}{'libsodium'}.
}
\details{
Please see \code{\link{sessionCookie}} for more information on how to save the generated key.
}
\seealso{
\code{\link{sessionCookie}}
}
