% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer.R
\name{register_serializer}
\alias{register_serializer}
\alias{registered_serializers}
\title{Register a Serializer}
\usage{
register_serializer(name, serializer, verbose = TRUE)

registered_serializers()
}
\arguments{
\item{name}{The name of the serializer (character string)}

\item{serializer}{The serializer to be added.}

\item{verbose}{Logical value which determines if a message should be printed when overwriting serializers}
}
\description{
A serializer is responsible for translating a generated R value into output
that a remote user can understand. For instance, the \code{serializer_json}
serializes R objects into JSON before returning them to the user. The list of
available serializers in plumber is global.
}
\section{Functions}{
\itemize{
\item \code{register_serializer}: Register a serializer with a name

\item \code{registered_serializers}: Return a list of all registered serializers
}}

