% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{register_ui}
\alias{register_ui}
\alias{registered_ui}
\title{Add UI for plumber to use}
\usage{
register_ui(ui)

registered_ui()
}
\arguments{
\item{ui}{A list of that plumber can use to mount
a UI.}
}
\description{
Add UI for plumber to use
}
\details{
\code{\link[=register_ui]{register_ui()}} is used by other packages like \code{swagger}.
When you load these packages, it calls \code{\link[=register_ui]{register_ui()}} to provide a user
interface that can interpret your plumber OpenAPI Specifications.

\code{ui} list expects the following values
\describe{
\item{package}{Name of the package required for the UI.}
\item{name}{Name of the UI.}
\item{index}{A function that returns the HTML content of the landing page of the UI.}
\item{static}{A function that returns the path to the assets (images, javascript, css, fonts) the UI will use.}
}
}
