% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer-json.R, R/serializer-content-type.R,
%   R/serializer-html.R, R/serializer-htmlwidget.R, R/serializer.R
\name{serializer_json}
\alias{serializer_json}
\alias{serializer_unboxed_json}
\alias{serializer_content_type}
\alias{serializer_html}
\alias{serializer_htmlwidget}
\alias{serializers}
\title{Plumber Serializers}
\usage{
serializer_json(na = "null")

serializer_unboxed_json(na = "null")

serializer_content_type(type)

serializer_html()

serializer_htmlwidget()
}
\arguments{
\item{type}{The value to provide for the \code{Content-Type} HTTP header.}
\item{na}{How should NA values be encoded?}

}
\description{
Serializers are used in Plumber to transform the R object produced by a
filter/endpoint into an HTTP response that can be returned to the client. See
\href{https://book.rplumber.io/rendering-and-output.html#serializers}{here} for
more details on Plumber serializers and how to customize their behavior.
}
