% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer-json.R, R/serializer-content-type.R,
%   R/serializer-csv.R, R/serializer-html.R, R/serializer-htmlwidget.R,
%   R/serializer-rds.R, R/serializer-yaml.R, R/serializer.R
\name{serializer_json}
\alias{serializer_json}
\alias{serializer_unboxed_json}
\alias{serializer_content_type}
\alias{serializer_csv}
\alias{serializer_html}
\alias{serializer_htmlwidget}
\alias{serializer_rds}
\alias{serializer_yaml}
\alias{serializers}
\title{Plumber Serializers}
\usage{
serializer_json(...)

serializer_unboxed_json(auto_unbox = TRUE, ...)

serializer_content_type(type)

serializer_csv(...)

serializer_html()

serializer_htmlwidget(...)

serializer_rds(version = "2", ascii = FALSE, ...)

serializer_yaml(...)
}
\arguments{
\item{...}{extra arguments supplied to respective internal serialization function.}

\item{auto_unbox}{automatically \code{\link[jsonlite]{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link[jsonlite]{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{type}{The value to provide for the \code{Content-Type} HTTP header.}

\item{version}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3). The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}

\item{ascii}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default) a binary one.
    See also the comments in the help for \code{\link[base]{save}}.}
}
\description{
Serializers are used in Plumber to transform the R object produced by a
filter/endpoint into an HTTP response that can be returned to the client. See
\href{https://book.rplumber.io/rendering-and-output.html#serializers}{here} for
more details on Plumber serializers and how to customize their behavior.
}
