% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_api_spec.R
\name{validate_api_spec}
\alias{validate_api_spec}
\title{Validate OpenAPI Spec}
\usage{
validate_api_spec(pr, verbose = TRUE)
}
\arguments{
\item{pr}{A Plumber API}

\item{verbose}{Logical that determines if a "is valid" statement is displayed. Defaults to \code{TRUE}}
}
\description{
Validate an OpenAPI Spec using \href{https://github.com/APIDevTools/swagger-cli}{Swagger CLI} which calls \href{https://github.com/APIDevTools/swagger-parser}{Swagger Parser}.
}
\details{
If the api is deemed invalid, an error will be thrown.

This function is VERY \lifecycle{experimental} and may be altered, changed, or removed in the future.
}
\examples{
\dontrun{
pr <- plumb_api("plumber", "01-append")
validate_api_spec(pr)
}
}
